refactor: review feedbacks
